A.html
You can run this code in VScode and see it using the show preview option on VScode. After viewing this you can touch any of the given links and it will redirect you to a webpage on your browser where you can see it full screen and again by touching the link it will redirect you to the certificates.